compositor: Make repaint loop exit frequently in test mode
When in test mode the clock is only updated by client intervention.  If
an animation is trying to run with a clock that won't advance the
compositor will remain stuck in a tight loop.

To stop this, we don't let weston_output_reschedule_repaint set
repaint_needed if we're already scheduled in test mode.