zoom: Only use a motion listener if a pointer is available
If no pointer is available for a seat, we no longer try to install a
motion listener.

Also, we now add a motion listener when a pointer device is plugged in.

Note that a seat may have a pointer when the device to move it has been
removed - in this case we still add the listener and it will work if
a device is plugged in to control the pointer.