zoom: Change how we select a zoom location
Any time the zoom is updated the best zoom target will be calculated
with the following priority:

1) The pointer that initiated the zoom. (even if that pointer has been
unplugged, its last known location is used)

2) The center of the current keyboard focus.  If the focus is split across
heads, multiple heads will zoom.

3) The center of the output that previously had a keyboard focus, if that
focus has disappeared.

4) The center of the first output.

It should be possible in all cases to zoom out with a keyboard binding,
even if the previous target is no longer available.  Once zoomed out
completely the zoom deactivates and it may no longer be possible to
zoom back in at that location.

The zoom may jump when the target changes.