rdp: Move heads to pending list properly
wl_list_insert_list() should be used for this, otherwise an empty list
becomes a loop. Currently we never see an empty list here, so it won't
break yet, but some upcoming changes will hit this.

Also gratuitously move the head_list list_init because it's conceptually
part of "moving" all the elements out of it into the pending list.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>